--- swift-starter-week1/Week1 [STEP 1] yy-ss99
@@ -1,4 +1,0 @@
-코드 스타터 캠프 1주차 미션을 위한 저장소입니다.
-1. yy-ss99
-2. 안녕하세요. 7기 코드 스타터에 참여하게 된 yy-ss99입니다. 닉네임이 이니셜이라서 그냥 **와이**라고 불러주세요.
-3. 잘 부탁드립니다 :)
